Moved Kusto datasource and supporting classes to separate directory.
