+++ src/Microsoft.Kusto.ServiceLayer/DataSource/Kusto/IKustoClient.cs
@@ -3,7 +3,7 @@ using System.Data;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Microsoft.Kusto.ServiceLayer.DataSource
+namespace Microsoft.Kusto.ServiceLayer.DataSource.Kusto
 {
     public interface IKustoClient
     {
